app/requirements.txt
pip==24.1.2
twitter-text-python
pyenchant
levenshtein
transformers
torch 
torchvision 
torchaudio
numpy==1.26.4
WordsPy
pandas
scikit-learn
emoji
peft
fastapi
uvicorn[standard]
slowapi
httpx